Fix all clippy warnings across the workspace, lint all crates
- Makefile clippy target now lints all 8 kernel/OS packages explicitly
  (not just zcore), using --no-deps to skip third-party crates
- Fix warnings in: zcore-drivers (clamp, transmute annotations,
  len_zero), kernel-hal (empty doc line, byte grouping, is_multiple_of,
  mut_from_ref allow, unnecessary casts, Default impl), zircon-object
  (derivable Default impls, unnecessary casts, for_kv_map, get_first,
  is_multiple_of, empty docs, doc continuation, new_without_default,
  multiple_bound_locations crate allow), zircon-syscall (unexpected_cfgs
  allow, never type fallback, legacy constants, contains, empty docs,
  unnecessary casts, transmute annotation, is_multiple_of, unit struct
  default), linux-object (multiple_bound_locations + double_must_use
  crate allows, derivable Default impls, for_kv_map, len_zero,
  unwrap_or_default, unnecessary_get_then_check, size_of_val,
  byte_char_slices, needless_as_bytes, result_unit_err allow),
  linux-syscall (repr packed ABI, div_ceil, slow_vector_initialization,
  unnecessary casts, needless_return, doc continuations, is_multiple_of),
  zcore-loader (inspect_err)
- Address code review: return ENOSYS for cross-process getpgid/getsid

+++ linux-syscall/src/lib.rs
@@ -254,13 +254,13 @@ impl Syscall<'_> {
             Sys::GETSID => self.sys_getsid(a0),
             Sys::GETPGID => self.sys_getpgid(a0),
             Sys::GETGROUPS => self.sys_getgroups(a0 as i32, a1.into()),
-            Sys::SETGROUPS => self.sys_setgroups(a0 as usize, a1.into()),
+            Sys::SETGROUPS => self.sys_setgroups(a0, a1.into()),
             Sys::SETPRIORITY => Ok(0), // scheduling priority — stub
             Sys::PRCTL => self.sys_prctl(a0, a1),
             Sys::MEMBARRIER => Ok(0), // memory barrier — no-op on single CPU
             Sys::PRLIMIT64 => self.sys_prlimit64(a0, a1, a2.into(), a3.into()),
             Sys::REBOOT => self.sys_reboot(a0 as u32, a1 as u32, a2 as u32),
-            Sys::GETRANDOM => self.sys_getrandom(a0.into(), a1 as usize, a2 as u32),
+            Sys::GETRANDOM => self.sys_getrandom(a0.into(), a1, a2 as u32),
             Sys::RT_SIGQUEUEINFO => self.sys_rt_sigqueueinfo(a0 as _, a1, a2.into()),
 
             // kernel module — not applicable for zCore
