+++ linux-syscall/src/time.rs
@@ -283,7 +283,7 @@ impl Syscall<'_> {
         }
 
         info!("tick: {:?}", tick);
-        Ok(tick as usize)
+        Ok(tick)
     }
 
     /// clock nanosleep
